### 2024-11-12 #7454 @editorialbot
[PRE REVIEW]: SynthEddy: A Python program for generating synthetic turbulent flow fields
**Submitting author:** @omltcat (<a href="http://orcid.org/0009-0009-1049-268X">Ange (Phil) Du</a>)
**Repository:** https://github.com/omltcat/turbulent-flow
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
