### 2024-11-16 038e165
Add poster from Nov 2024 presentation
