### 2024-11-16 0ab8a49
Change visualize color map to jet to be more inline with other CFD software outputs (used in poster)
