### 2025-04-10 20c2b2a
Updates in-code docs for thesis. Adds code churn scripts
