+++ gitcodechurn.py
@@ -1,0 +1,315 @@
+'''
+Authors: 
+- Francis Laclé
+- Jonathan Guerne
+License: MIT
+Version: 1.2.1
+
+Script to compute "true" code churn of a Git repository.
+
+Code churn has several definitions, the one that to me provides the
+most value as a metric is:
+
+"Code churn is when an engineer rewrites their own code in a short time period."
+
+Reference: https://blog.gitprime.com/why-code-churn-matters/
+
+This lightweight script looks at commits per author for a given date range on
+the default branch. For each commit it bookkeeps the files that were changed
+along with the lines of code (LOC) for each file. LOC are kept in a sparse
+structure and changes per LOC are taken into account as the program loops. When
+a change to the same LOC is detected it updates this separately to bookkeep the
+true code churn.
+
+Result is a print with aggregated contribution and churn per author for a
+given time period.
+
+Tested with Python version 3.5.3 and Git version 2.20.1
+
+'''
+
+import subprocess
+import shlex
+import os
+import argparse
+import datetime
+from time import strptime
+
+def main():
+    parser = argparse.ArgumentParser(
+        description = 'Compute true git code churn to understand tech debt.',
+        usage       = 'python [*/]gitcodechurn.py after="YYYY[-MM[-DD]]" before="YYYY[-MM[-DD]]" author="flacle" dir="[*/]path" [-exdir="[*/]path"]',
+        epilog      = 'Feel free to fork at or contribute on: https://github.com/flacle/truegitcodechurn'
+    )
+    parser.add_argument(
+        'after',
+        type = str,
+        help = 'search after a certain date, in YYYY[-MM[-DD]] format'
+    )
+    parser.add_argument(
+        'before',
+        type = str,
+        help = 'search before a certain date, in YYYY[-MM[-DD]] format'
+    )
+    parser.add_argument(
+        '-start-commit',
+        type = str,
+        help = 'search from a certain commit, in short SHA format'
+    )
+    parser.add_argument(
+        '-end-commit',
+        type = str,
+        help = 'search to a certain commit, in short SHA format'
+    )
+    parser.add_argument(
+        'author',
+        type = str,
+        help = 'an author (non-committer), leave blank to scope all authors'
+    )
+    parser.add_argument(
+        'dir',
+        type = dir_path,
+        default = '',
+        help = 'the Git repository root directory to be included'
+    )
+    parser.add_argument(
+        '-exdir',
+        metavar='',
+        type = str,
+        default = '',
+        help = 'the Git repository subdirectory to be excluded'
+    )
+    args = parser.parse_args()
+
+    after  = args.after
+    before = args.before
+    start_commit = args.start_commit
+    end_commit = args.end_commit
+    author = args.author
+    dir    = args.dir
+    # exdir is optional
+    exdir  = args.exdir
+
+    # for the positionals we remove the prefixes
+    # TODO not sure why this is happening
+    after   = remove_prefix(after, 'after=')
+    before  = remove_prefix(before, 'before=')
+    author  = remove_prefix(author, 'author=')
+    # dir is already handled in dir_path()
+
+    # retrive start and end commit date from given commit hashes
+    start_commit_date = get_commit_timestamp(start_commit, dir) if start_commit else None
+    end_commit_date = get_commit_timestamp(end_commit, dir) if end_commit else None
+
+    # sort commit date if necessary
+    if start_commit_date and end_commit_date:
+        start_commit_date, end_commit_date = sorted([start_commit_date, end_commit_date])
+
+    # return specified commit if it is the same for the start and end value
+    # otherwise look for commits in between
[…]
