### 2025-04-10 3179500
Shortens the paper.
