### 2025-04-13 #7454 @arfon
@baperry2 @yueyanglu @sthavishtha :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is SynthEddy: A Python program for generating synthetic turbulent flow fields

Many thanks
Arfon
