### 2025-04-14 #7454 @arfon
@mliza – 👋 would you be willing to review this submission for JOSS? The submission under consideration is SynthEddy: A Python program for generating synthetic turbulent flow fields

Many thanks
Arfon
