### 2025-06-26 omltcat/turbulent-flow#84 @baperry2
JOSS Review - Documentation - Community Guidelines are incomplete
The files `CONTRIBUTING.md` and `CodeOfConduct.md` are currently placeholders. These should be replaced with specific instructions for contributing to, reporting issues with, and seeking support for this software.
