### 2025-06-26 omltcat/turbulent-flow#85 @baperry2
JOSS Review - Software Paper
The following items should be addressed in the paper:

- [ ] Statement of Need: Does not adequately contextualize the current tool relative to alternative strategies/software
- [ ] References: The references for general use of LES seem to be somewhat arbitrarily selected, but i suppose that's ok. But it's weird to have several references for general use of LES but only one for synthetic turbulence field generation, which is the focus of this work. Additional references contextualizing other turbulent flow field generation strategies/software should be added.
- [ ] Quality of Writing: While the writing is overall ok, there are several grammatical errors that should be corrected. For example, the last line of pg 1 should be "The user can generate a turbulent velocity field and query it for use in both turbulence research and as an IC/BC for CFD simulations.". The paragraph beginning with "The generated field is fully wrapped..." is difficult to understand and needs to be revised for clarity. Beyond these specific examples, I'd recommend a full read through by the authors to check for grammatical errors.
- [ ] Formatting: Bulleted lists in the "Document driven development" and "Customization" sections are not being rendered properly.
