### 2025-06-26 #8086 @baperry2
My review is in progress. I'll use this comment to track issues I've created for the review.

### General Checks

None so far

### Functionality

None so far

### Documentation

- **Community Guidelines** https://github.com/omltcat/turbulent-flow/issues/84

### Software Paper

- Relating to all unchecked items in this section: https://github.com/omltcat/turbulent-flow/issues/85
