### 2025-06-27 3692075
Fix Makefile conda activation under Linux and MacOS. See #86
