### 2025-07-14 #8086 @mliza
My review so far:

**Paper**
“Turbulent flow is a type of fluid motion characterized by chaotic fluctuations in velocity and pressure.” Look at Pope, Davidson and Tennekes (specially Tennekes). He provides a better characterization of turbulent flows (Irregularity, Diffusivity, Large Reynolds Numbers, 3D, Dissipation, Continuum). Turbulent flows also have fluctuations on temperature, density, etc.

“LES can capture more intricate details of a turbulent flow, down to individual eddies” (I would add something up to where, what do you mean by individual eddies? In general LES will capture a percentage of the cascade ~80% and the rest will be modeled).   Is your tool made for incompressible or compressible turbulence? (I am assuming incompressible), but add this somewhere and be specific about this.

For Docs I will suggest using Read the docs, here are a few examples (https://haot.readthedocs.io/en/latest/, https://ambiance.readthedocs.io/en/stable/, ). People like pretty plots and colorfule figures.

**REPO**
At the top level you are missing:
- CHANGELOG.md
- I suggest using git tags for versioning (https://dev.to/neshaz/a-tutorial-for-tagging-releases-in-git-147e)
- A lot of death code .gitmodules, .gitignore are empty files
- Contributing.md and CodeOfConduct has no information. Remove it or update it
-  Lack of documentation, most of the documentation files are empty

Unfortunalty I am not able to install the pacakge for proceed with the rest of my review.  Here are some suggestions:
- Use PyPI (not everyone uses conda)
- Use a pyproject.toml, so I can just clone your repo and do pip install -e .
- Provide instructions on how to install it if you dont have and can not install conda in your computer
